♻️[Refactor] 파일 분리, 불필요한 주석 지움
